use HudRenderCallback.EVENT instead of InGameHud#renderStatusEffectOverlay
