+++ src/main/resources/offershud.mixins.json
@@ -5,7 +5,6 @@
     "compatibilityLevel": "JAVA_17",
     "mixins": [],
     "client": [
-        "InGameHudMixin",
         "ReceiveTradeOfferPacket",
         "ValidTrade",
         "CloseScreen",
